Added a function to tell whether iFolder is running or not.
Also added code to use EEL to popup a confirmation dialog box
when reverting an iFolder.  Also added code to be able to popup
error messages to the user and cause the dialog to be modal on
the nautilus window.  All the error messages are not integrated,
but the capability is there and works.


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@2762 db2431b9-df06-0410-971c-a38ea9a75485