added code to fix https://bugzilla.novell.com/show_bug.cgi?id=73332
This has been fixed by changing the sync events.

The CollectionSyncEventArgs has a Connected instead of a Successful property.
To know if the sync is complete you will need to call SyncClient.GetCountToSync

This CollectionSyncEvent is now generated up to three times during a sync cycle.
1 )once with the Action set to StartLocalSync (3) which means we are syncing with the local file system. This is a new Action type.
2 ) if we can connect to the server with the Action set to StartSync (1) which means that we are sync-ing with the server.
3) one with the Action set to StopSync (2) which means we are done.

FileSyncEventArgs has also changed a new Direction of Local (3) has been added.
This event will be generated when we detect local file system changes.


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk/ifolder@4450 db2431b9-df06-0410-971c-a38ea9a75485