Fix for:
Bug 81808 - Novell.iFolder.iFolderApplication:OniFolderChangedEvent - System.NullReferenceException: Object reference not set to an instance of an object

Bug 78583 - Linux Client crashed when attempting to run overnight test

In some cases, we are being called with an event on an iFolder that we don't have an object for in the iFolderData instance.  We were just assuming that we could get an iFolderHolder object for it and weren't checking for a null case.  This fix checks for null before just letting the code run wild (or blow up, as it were).


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@4559 db2431b9-df06-0410-971c-a38ea9a75485