Fix for Bug 73324 - Client fails to authenticate if LDAP username an international character in it.
We were encoding the request with the default encoding.  In NLD the default is UTF-8.  Mono (BasicClient.cs I believe) just chops off the extra bytes of UTF-8, essentially making the request send 8859.  The server attempted to decode the username with UTF-8.  I've now changed the code to force us to encode in 8859 (Windows-1252 codepage) which allows the user to login fine.  Kudos to mlasky for helping out with this!


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@4914 db2431b9-df06-0410-971c-a38ea9a75485