Fix for Bug 83868 - Account dialog displays incorrect status.
This is the first checkin into HEAD of the "Event-based" domains for the Linux Client.  The only window that uses this so far is the Accounts Page in iFolder Preferences, but other dialogs were affected.  The Account Details dialog still needs to be implemented to support these events, but is not necessary to fix this bug.


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@4981 db2431b9-df06-0410-971c-a38ea9a75485