Fix for Bug 120007 - User prompted for password for accounts that already have the password saved.
There were a couple places where the OnFieldsChanged and AccSelectionChangedHandler events were being called when they shouldn't have been.  They were clearing out the password at the wrong times.  One of the times was when an existing domain (with password saved) was highlighted and then the user pressed the "Add" button.  The other time was when multiple domains existed in the list and the user moved from one that had the password saved to another account.

File changed:
	src/LinuxClient/library/PrefsAccountsPage.cs


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/R3OESSP2RC1_BR@5351 db2431b9-df06-0410-971c-a38ea9a75485