Fix for Bug 129608 - Double-clicking on user object in Sharing UI should open up the access dialog UI
You've always been able to right-click over the "Access" column to pop up a little menu to change a user's rights.  I've made it so that if you right-click anywhere on the row (in different columns) it will pop open the menu (just like the Mac Client already does).I've also modified the behavior so that when you double-click on a row, it will perform the same action as the "Access" button.

File modified (in 3.2 Branch and CVS Head):
	src/LinuxClient/library/iFolderPropSharingPage.cs


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/R3OESSP2RC1_BR@5454 db2431b9-df06-0410-971c-a38ea9a75485