Fix for a usability problem during "Revert and Delete" task.
Added code to fix one of the common problems users ran into during usability testing.  Users were asked to remove their iFolder
from the server.  It was unclear to them when they'd highlight it and revert...and then it still appeared on the server.  The
fix provided by this check-in adds a checkbox to the revert dialog that says, "Also delete this iFolder from the server" so users
are trained about what happens when reverting (i.e., the iFolder will still be available on the server).  It also allows users
to revert and delete all at once.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4/ifolder@5855 db2431b9-df06-0410-971c-a38ea9a75485