Fix for Bug 141941 - New client doesn't display all ifolder 'Name or Path'.
I added a bunch of calls to GLib.Markup.EscapeText() in all the places where we use Markup Text.  The real problem was happening
because of an iFolder Name/Path that had an ampersand character followed by a space.  Pango Markup would attempt to interpret
this as an entity and fail.  Calling GLib.MarkupEscapeText() fixes the problem by changing '&' to "&amp;".

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4/ifolder@5858 db2431b9-df06-0410-971c-a38ea9a75485