Fix for Bug 148955 - Automatically connect when iFolder is started value is not respected.
Renamed the checkbox string to, "Automatically connect" since the connecting will be attempted at the beginning of every
synchronization cycle and not just when iFolder starts.

It is valid for an account to be connected and then for the user to uncheck the "Automatically connect" checkbox.  If this occurs,
the account should still be connected and synchronize.  If, however, the user then manually logs out of the server, it will not
automatically reconnect during the next synchronization cycle.

I also fixed up a problem in Simias/src/DomainWatcher/DomainWatcher.cs to respect the setting.  It was inadvertently omitted in
an earlier check-in.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@5933 db2431b9-df06-0410-971c-a38ea9a75485