Fix for Bug 135594 - iFolder3 killed my nautilus.
Fixed in 3.4 branch and svn trunk.

Added mutexes around the GLib Idle IDs for reconnect and disconnect.  This should guarantee that we don't hit a race condition and also cut down on high CPU utilization on a g_idle_add() that is called over and over under the right circumstances.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4@6047 db2431b9-df06-0410-971c-a38ea9a75485