Fix for Bug 161970 - iFolder 3.4 - Reverting iFolder during first "Checking for Changes" causes simias events to be lost.
The problem was that the client was not catching exceptions if they occurred while processing an event from Simias.  If this happens, the event server in simias disconnects the offending client so it doesn't have to deal with processing it anymore.

I added try/catch statements in all of the places that process simias events to guarantee we never throw an exception.  I also had to change code to make sure we don't call ListStore.EmitRowChanged() when the TreeIter is no longer valid.

Additionally, there's a new GConf setting "/apps/ifolder3/debug/print_simias_event_errors" that when set to TRUE will allow exceptions to be printed to the console.  The default setting for this is FALSE.  While adding this, I fixed a problem where the ClientConfig was looking for event notification using the "/apps/ifolder" instead of "/apps/ifolder3".

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4@6098 db2431b9-df06-0410-971c-a38ea9a75485