Fix for Bug 158916 - iFolder 3.4 Client unloads after setting up account - Backtrace was generated from '/usr/bin/ifolder'.
DomainController.cs  - Changed all the DomainAdded events so that they are guaranteed to call other event handlers on the GUI thread.
iFolderIconView.cs   - Fixed a problem where we were attempting to add iFolderViewGroup (the server group in the main window) multiple times.
iFolderWaitDialog.cs - Fixed a problem where progressbar.pulse() was still being called after the dialog was no longer visible.

I wasn't able to reproduce this bug, but the above fixes all caused Gtk-CRITICAL error messages to be logged to the console.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4/ifolder@6101 db2431b9-df06-0410-971c-a38ea9a75485