Bug 160421 - iFolder 3.4 - Novell.iFolder.iFolderApplication:OnReLoginDialogResponse exception causes client to unload.
Added code in Novell.iFolder.iFolderApplication:OnReLoginDialogResponse() to check to see if the domain/account is still present.  If it's not it will pop up an error message to the user letting them know it no longer exists and instructs them to add it in the Account Settings Dialog if they want to connect to it again.  If an exception is thrown during the login process, a try/catch prevents a blowup and shows an error to the user.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4@6105 db2431b9-df06-0410-971c-a38ea9a75485