Partial fix for Bug 165068 - iFolder 3.4 - Changing the address in the Server Property can get into state where user can no longer use account.
I found a few extra bugs related to this.
1. The event listeners on DomainController were not being deregistered for and so if you opened
the dialog again, you'd be called multiple times when events on the domain occured (like the
host name changing).
2. The DomainController code was not checking the return value of the method to set the host address.
3. Added better error handling in DomainController to check the value SimiasWebService returns to
validate there really was a change.
4. Added better error handling in the details dialog when an error/exception occurs when changing the
server host address.

The real problem still is that the SimiasWebService doesn't actually change the host address even
though it reports success.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4/ifolder@6173 db2431b9-df06-0410-971c-a38ea9a75485