Fix for Bug 167655 - iFolder 3.4 - Help fails to come up.
I'm not sure why, but Catherine renamed "front.html" -> "bookinfo.html" so our help files weren't loading correctly.  I modified the code to reference a static variable in Util.cs for the main page so that if this happens again, it will be easy to switch all references.  It now uses "bookinfo.html" to load the main page of the help documentation.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4/ifolder@6196 db2431b9-df06-0410-971c-a38ea9a75485