I removed code from the Linux client that was a workaround for when the SimiasWebService.SetDomainHostAddress() wasn't working correctly.  Now that I've removed the workaround, I can:
- Change "193.97.125.1" -> "193.97.125.1:80" successfully
- Change "193.97.125.1:80" -> "193.97.125.1" successfully
- Change "193.97.125.1" -> "193.97.125.1:1234" successfully
- Change "193.97.125.1:1234" -> "193.97.125.1" successfully

Note: Bruce could likely shed a bit more light on this, but, when I change "193.97.125.1" -> "193.97.125.1:443", the call actually goes through correctly, but the Host returned by the domain does not show 443.  The user is still able to log in correctly to the server after changing to use 443.  I believe this may be something to do with 443 being the default port.

Also, after changing to use "193.97.125.1:80", I am not able to log in to the server.  This is expected since this server requires HTTPS (443) under the covers.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/branches/ifolder_3_4/ifolder@6197 db2431b9-df06-0410-971c-a38ea9a75485