Fix for Bug 182238 - Nautilus - Nautilus crashes when user attempts to create an iFolder on a mounted samba share.
The default return value to "can_be_ifolder" was set to be TRUE.  get_file_path() returns NULL on the files inside of remotely connected file systems.  So, when this was the case the can_be_ifolder() would incorrectly report that the folder could be converted to an iFolder.  I modified the default return value of can_be_ifolder() to be FALSE.

git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@6433 db2431b9-df06-0410-971c-a38ea9a75485